+++ keanu-project/src/main/java/io/improbable/keanu/tensor/TensorValidator.java
@@ -1,6 +1,7 @@
-package io.improbable.keanu.tensor.dbl;
+package io.improbable.keanu.tensor;
 
 import io.improbable.keanu.tensor.bool.BooleanTensor;
+import io.improbable.keanu.tensor.dbl.DoubleTensor;
 
 public interface TensorValidator {
     BooleanTensor check(DoubleTensor tensor);
