+++ keanu-project/src/main/java/io/improbable/keanu/tensor/TensorValueEqualsValidator.java
@@ -1,6 +1,7 @@
-package io.improbable.keanu.tensor.dbl;
+package io.improbable.keanu.tensor;
 
 import io.improbable.keanu.tensor.bool.BooleanTensor;
+import io.improbable.keanu.tensor.dbl.DoubleTensor;
 
 public class TensorValueEqualsValidator implements TensorValidator {
